+++ wagon_sync/process_code.py
@@ -6,15 +6,7 @@ from wagon_sync.code_edition import replace_content
 from wagon_sync.params.delimiters import (
     CHALLENGIFY_DELIMITERS,
     CHALLENGIFY_REPLACEMENTS,
-    # meta delimiters
-    META_DELIMITER_VERSION_REPLACEMENT,
-    META_DELIMITER_BEFORE_BEGIN,
-    META_DELIMITER_BEFORE_END,
-    META_DELIMITER_ONLY_BEGIN,
-    META_DELIMITER_ONLY_END,
-    META_DELIMITER_AFTER_BEGIN,
-    META_DELIMITER_AFTER_END,
-)
+    ITERATE_DELIMITERS)
 
 
 def process_delimiters(content, file_extension, verb_delimiters):
@@ -44,6 +36,34 @@ def process_delimiters(content, file_extension, verb_delimiters):
     return content
 
 
+def process_versions(content, rule, versions, keep=True):
+
+    # retrieve delimiter patterns
+    version_delimiter_begin = ITERATE_DELIMITERS[rule]["begin"]
+    version_delimiter_end = ITERATE_DELIMITERS[rule]["end"]
+
+    # iterate through versions
+    for version in versions:
+
+        # replace versions in delimiters
+        delimiter_begin = version_delimiter_begin.replace("version", version)
+        delimiter_end = version_delimiter_end.replace("version", version)
+
+        # handle content
+        if keep:
+
+            # remove delimiters
+            content = content.replace(delimiter_begin, "")
+            content = content.replace(delimiter_end, "")
+
+        else:
+
+            # remove block
+            content = replace_content(content, "", delimiter_begin, delimiter_end)
+
+    return content
+
+
 def process_code(source, destination, file_extension, version_iterator=None):
 
     # create destination directory
@@ -51,74 +71,34 @@ def process_code(source, destination, file_extension, version_iterator=None):
 
     # read content
     with open(source, "r") as file:
-        source_content = file.read()
+        content = file.read()
 
     # run challengify
     if version_iterator is None:
 
         # process content through challengify delimiters
-        replaced_content = process_delimiters(source_content, file_extension, CHALLENGIFY_DELIMITERS)
+        content = process_delimiters(content, file_extension, CHALLENGIFY_DELIMITERS)
 
     # run challengify iterate
     else:
 
-        # retrieve current version for delimiters
-        challenge_position = version_iterator.iterated_position
+        # retrieve versions
+        versions_before_current = version_iterator.get_versions_before()
+        versions_after_current = version_iterator.get_versions_after()
+        version_current = version_iterator.get_version_current()  # remove usage of version_iterator.iterated_position
 
-        # iterate through meta delimiters
-        for delimiter_version in version_iterator.versions:  # iterate through all versions without using the iterator
+        # process only to
+        content = process_versions(content, "only_to", versions_before_current, keep=False)
+        content = process_versions(content, "only_to", versions_after_current + version_current)
 
-            # retrieve challenge versions for delimiters
-            meta_version_position = delimiter_version.position
-            meta_version_name = delimiter_version.version
+        # process only for
+        content = process_versions(content, "only_for", versions_before_current + versions_after_current, keep=False)
+        content = process_versions(content, "only_for", version_current)
 
-            # build meta version delimiters
-            meta_before_begin = META_DELIMITER_BEFORE_BEGIN.replace(META_DELIMITER_VERSION_REPLACEMENT, meta_version_name)
-            meta_before_end = META_DELIMITER_BEFORE_END.replace(META_DELIMITER_VERSION_REPLACEMENT, meta_version_name)
-            meta_only_begin = META_DELIMITER_ONLY_BEGIN.replace(META_DELIMITER_VERSION_REPLACEMENT, meta_version_name)
-            meta_only_end = META_DELIMITER_ONLY_END.replace(META_DELIMITER_VERSION_REPLACEMENT, meta_version_name)
-            meta_after_begin = META_DELIMITER_AFTER_BEGIN.replace(META_DELIMITER_VERSION_REPLACEMENT, meta_version_name)
-            meta_after_end = META_DELIMITER_AFTER_END.replace(META_DELIMITER_VERSION_REPLACEMENT, meta_version_name)
-
-            # version x removes content with meta delimiters before x and down
-            if challenge_position >= meta_version_position:
-
-                # replace meta delimiters outside of version number by DELETE delimiters (remove content)
-                source_content = source_content.replace(meta_before_begin, RAW_CODE_DELETE_BEGIN)
-                source_content = source_content.replace(meta_before_end, RAW_CODE_DELETE_END)
-
-            else:
-
-                # remove delimiters inside of version number (keep content)
-                source_content = source_content.replace(meta_before_begin, "")
-                source_content = source_content.replace(meta_before_end, "")
-
-            # version x removes content with meta delimiters if not equal to x
-            if challenge_position != meta_version_position:
-
-                # replace meta delimiters outside of version number by DELETE delimiters (remove content)
-                source_content = source_content.replace(meta_only_begin, RAW_CODE_DELETE_BEGIN)
-                source_content = source_content.replace(meta_only_end, RAW_CODE_DELETE_END)
-
-            else:
-
-                # remove delimiters inside of version number (keep content)
-                source_content = source_content.replace(meta_only_begin, "")
-                source_content = source_content.replace(meta_only_end, "")
-
-            # version x removes content with meta delimiters after x and up
-            if challenge_position <= meta_version_position:
-
-                # replace meta delimiters outside of version number by DELETE delimiters (remove content)
-                source_content = source_content.replace(meta_after_begin, RAW_CODE_DELETE_BEGIN)
-                source_content = source_content.replace(meta_after_end, RAW_CODE_DELETE_END)
-
-            else:
-
-                # remove delimiters inside of version number (keep content)
-                source_content = source_content.replace(meta_after_begin, "")
-                source_content = source_content.replace(meta_after_end, "")
+        # process only from
+        content = process_versions(content, "only_from", versions_after_current, keep=False)
+        content = process_versions(content, "only_from", versions_before_current + version_current)
 
     # write content
     with open(destination, "w") as file:
-        file.write(replaced_content)
+        file.write(content)